ExpressionCapturer tests WIP
