+++ src/Fluent.Calculations.Primitives/Expressions/Capture/IExpressionMembersCapturer.cs
@@ -1,7 +1,7 @@
 ﻿namespace Fluent.Calculations.Primitives.Expressions.Capture;
 using System.Linq.Expressions;
 
-internal interface IExpressionMembersCapturer
+public interface IExpressionMembersCapturer
 {
     List<object> Capture(Expression expression);
 }
